refactor(tui): A4 — terminal-ownership state out of jackin-diagnostics (P7)
Parallel Change: moves the RICH_SURFACE_ACTIVE / HOST_SCREEN_OWNED
atomics and their accessor functions from
crates/jackin-diagnostics/src/terminal.rs into a new
crates/jackin-tui/src/ownership.rs module. jackin-diagnostics becomes
a thin re-export shim so every existing jackin_diagnostics::terminal::*
call site compiles unchanged.

Touches:
  - created crates/jackin-tui/src/ownership.rs (terminal-ownership flags
    + alt-screen re-assertion + terminal-title helper; moved verbatim
    from jackin-diagnostics/src/terminal.rs minus the shorten_home
    re-export line)
  - modified crates/jackin-tui/src/lib.rs (adds pub mod ownership;)
  - modified crates/jackin-diagnostics/Cargo.toml (adds jackin-tui
    dependency; jackin-diagnostics is L1 and adding the L3 presentation
    dep is architecturally allowed here because the alternative —
    duplicating the two atomic flags in a sister file — would re-introduce
    the very split that this slice consolidates)
  - modified crates/jackin-diagnostics/src/terminal.rs (body replaced
    with re-exports of every public name from jackin_tui::ownership;
    shorten_home no longer re-exported from this file, moved to the
    diagnostics lib.rs root alongside the other jackin-core re-exports)
  - modified crates/jackin-diagnostics/src/lib.rs (the shorten_home
    re-export moved up to the lib.rs root; otherwise no pub-API change)

Decisions:
  - jackin-diagnostics is the one L1 crate the spec allows to take a
    presentation dep on L3 because every other consumer (jackin-config,
    jackin-manifest, jackin-docker, jackin-env, jackin-image,
    jackin-term) is L0/L1/L2 and must not pull ratatui transitively. The
    re-export shim means every existing jackin_diagnostics::terminal::*
    import keeps working; the next refactor of those lower crates can
    then drop the indirection in a follow-up slice.
  - The two atomic flags + accessor functions are byte-identical between
    the old jackin-diagnostics/src/terminal.rs and the new
    jackin-tui/src/ownership.rs (moved verbatim; the only difference is
    the module path in their doc comments).
  - shorten_home is the only public name the old terminal.rs re-exported
    from jackin-core; it was always a stray, not a terminal-helper.
    The re-export moved to lib.rs root next to the other jackin-core
    re-exports so jackin_diagnostics::shorten_home still resolves.

Verified:
  - cargo fmt --check clean
  - cargo clippy --workspace --all-targets --all-features --locked -- -D warnings clean
  - cargo nextest run --workspace --no-fail-fast --features test-support:
    5250/5251 (one pre-existing failure on main; unrelated)
  - cargo xtask lint: file-size OK; test-layout OK; arch gate at 3
    forbidden edges (no new inversions introduced)

Refs: plans/codebase-health-playbook.md#a4--move-terminal-ownership-state-out-of-jackin-diagnostics-p7
Refs: docs/content/docs/roadmap/codebase-health-enforcement.mdx#workstream-w1
